third_party/WebKit/LayoutTests/http/tests/devtools/tracing/tracing-timeline-load-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests tracing based Timeline save/load functionality.


Running: testNormal
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.processingStarted()
TimelineLoaderClient.loadingComplete(true)
Saved data is equal to restored data: true

Running: testJSONObjectFormat
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.processingStarted()
TimelineLoaderClient.loadingComplete(true)
Saved data is equal to restored data: true

Running: testJSONObjectFormatWithMetadata
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.processingStarted()
TimelineLoaderClient.loadingComplete(true)
Saved data is equal to restored data: true

Running: testBroken
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.processingStarted()
TimelineLoaderClient.loadingComplete(true)
Model is empty: true

Running: testMalformedJSON
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.loadingComplete(false)
Model is empty: true
